Split ServicePage content into separate tab
