+++ src/js/routes/services.js
@@ -3,6 +3,7 @@ let Route = Router.Route;
 
 import ServiceOverlay from '../components/ServiceOverlay';
 import ServicesPage from '../pages/ServicesPage';
+import ServicesTab from '../pages/services/ServicesTab';
 
 let servicesRoutes = {
   type: Route,
@@ -10,6 +11,11 @@ let servicesRoutes = {
   path: 'services/?',
   handler: ServicesPage,
   children: [
+    // Undocumented RR 0.13.5 behaviour: unnamed routes are treated as default
+    {
+      type: Route,
+      handler: ServicesTab
+    },
     {
       type: Route,
       name: 'services-detail',
